ANDROID: PROTOTYPE: Add hold functionality to schedtune CPU boost
When tasks come and go from a runqueue quickly, this can lead to boost
being applied and removed quickly which sometimes means we cannot raise
the CPU frequency again when we need to (due to the rate limit on
frequency updates). This has proved to be a particular issue for RT tasks
and alternative methods have been used in the past to work around it.

This is an attempt to solve the issue for all classes and cpufreq
governors by introducing a generic mechanism in schedtune to retain
the max boost level from task enqueue for a minimum period - defined
here as 50ms. This timeout was determined experimentally and is not
configurable.

It works like so:

When we enqueue a task, we store the time (in ns) at which we will
remove the boost associated with it. The shortest timeout in effect is
stored. If the task is active longer than the hold time, nothing
changes.

When we check CPU boost level, we also check the closest timeout. If
that has not passed yet, then we use the stored boost.

If that timeout is in the past, then we re-evaluate the boost groups for
that CPU, taking into account the timeouts associated with each group.
Again, the nearest one is stored at the top level.

Timeouts are set to zero when they are expired.

Idea based on rt-boost-retention patches from Joel.

Change-Id: I52cc2d2e82d1c5aa03550378c8836764f41630c1
Suggested-by: Joel Fernandes <joelaf@google.com>
Signed-off-by: Chris Redpath <chris.redpath@arm.com>