ANDROID: arm-smccc: backport SMCCC 1.1 definitions
This is a partial backport of 09e6be12effd ("arm/arm64: KVM: Advertise
SMCCC v1.1"), which is part of a large patchstack against KVM.

The KVM patchstack is a messy backport, and we don't support using
Android kernels to host KVM guests.  So just take the new ARM_SMCCC_*
constants that will be used later in the patchstack.

From the original commit:

    Tested-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
    Reviewed-by: Christoffer Dall <christoffer.dall@linaro.org>
    Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
    Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>

Change-Id: I2dcfaffa411563f95e56d2e6f4a3760dc00d025a
Signed-off-by: Greg Hackmann <ghackmann@google.com>