Update patch set 1
Patch Set 1: Code-Review-1

(13 comments)

Hi Chris, I've collected some of the considerations we have in our chat.

I'm of the idea that we can try to have a simpler solution based on max aggregation of the timeouts collected at enqueue time. Thus, touching only the fast path while completely ignoring updates from the slow path (when tasks migrate groups or we change boost value).

At the end such a solution should be really close to the code we already have to add +1 every time we enqueue a task. Where, instead of +1 we update the next timeout for a boostgroup based on the last enqueue task, and aggregate the "boost_timeout" the same way we already aggregate "boost_max"...

By focusing just on the fast path, we can also use sched_clock_cpu(), or directly the rq->clock

The -1 is just to flags the proposed solution as a possible alternative to investigate before merging this one.

Patch-set: 1
Reviewer: Gerrit User 1119068 <1119068@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1