Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Hi Dietmar,
> > 
> > So in my understanding, this is basically giving a strong priority to the capacity values coming from the energy model. But in the case where the max freq of the CPUs isn't known at compile time, it's more than likely that the scaled capacity values (computed with dmips-capacity-mhz) are actually more accurate than the capacity values of the EM no ?
> 
> I would also prefer a solution in which we let the arch_topology driver do his job, setting cpu_scale based on dmips-capacity-mhz dt values and max cpu frequency values from the cpufreq subsystem. But in this case, we have to make sure that the energy model we setup independently from the arch_topology driver matches the cpu_scale values.

Well, you can't just check that the cpu_scale values match. The max cap value in the energy model was measured on a specific device that has a specific max freq. If you boot another device with a different max freq, the values computed by the arch topology driver will be different from the ones in the EM, and it's not acceptable to just though an error here. What you really want is to either discard the capacity states of the EM that don't match available OPPs on the platform, or at the very least re-scale the capacity values in the EM to match the true capacities of the CPUs (which were computed by the arch topology driver).

Patch-set: 1
Reviewer: Gerrit User 1200908 <1200908@85c56323-6fa9-3386-8a01-6480fb634889>