Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > Hi Dietmar,
> > > 
> > > So in my understanding, this is basically giving a strong priority to the capacity values coming from the energy model. But in the case where the max freq of the CPUs isn't known at compile time, it's more than likely that the scaled capacity values (computed with dmips-capacity-mhz) are actually more accurate than the capacity values of the EM no ?
> > 
> > I would also prefer a solution in which we let the arch_topology driver do his job, setting cpu_scale based on dmips-capacity-mhz dt values and max cpu frequency values from the cpufreq subsystem. But in this case, we have to make sure that the energy model we setup independently from the arch_topology driver matches the cpu_scale values.
> 
> Well, you can't just check that the cpu_scale values match. The max cap value in the energy model was measured on a specific device that has a specific max freq. If you boot another device with a different max freq, the values computed by the arch topology driver will be different from the ones in the EM, and it's not acceptable to just though an error here. What you really want is to either discard the capacity states of the EM that don't match available OPPs on the platform, or at the very least re-scale the capacity values in the EM to match the true capacities of the CPUs (which were computed by the arch topology driver).

Just to clarify: By matching cpu_scale I meant that the cpu_scale value matches the capacity value of the highest element in the cap states vector.

I agree, but so far I was under the assumption that we could have different energy model dt data for platforms with different max frequency values.

Then if they don't match, this would be a configuration error and we could prevent EAS from starting or even the whole system.

If we want to have only _one_ energy model dt file for all those systems with different max frequency, then scaling the energy model capacity values would be an option. This has to be only done on the little cpus since big cpu capacity values are always scaled to 1024.

With the current solution (todays android-4.14 or even with this patch-set) changes in max frequency values (related to those which were there when the Energy Model was constructed) would go unnoticed.

Patch-set: 1
Reviewer: Gerrit User 1201522 <1201522@85c56323-6fa9-3386-8a01-6480fb634889>