BACKPORT: kasan, arm64: clean up KASAN_SHADOW_SCALE_SHIFT usage
(Commit 9597e74396c02c8189215dcdfc08088a9b8d8087 upstream).

This is a follow up patch to the series I sent recently that cleans up
KASAN_SHADOW_SCALE_SHIFT usage (which value was hardcoded and scattered
all over the code). This fixes the one place that I forgot to fix.

The change is purely aesthetical, instead of hardcoding the value for
KASAN_SHADOW_SCALE_SHIFT in arch/arm64/Makefile, an appropriate variable
is declared and used.

Signed-off-by: Andrey Konovalov <andreyknvl@google.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
Bug: 128674696
Test: build, boot and insmod test_kasan.ko with various configs
Signed-off-by: Andrey Konovalov <andreyknvl@google.com>