Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review-1
> 
> Any specific reason why this should not just be submitted to the 4.4.y stable kernel update upstream instead of putting it only in here in android-common?

No good reason. I was under the assumption that PELT in 4.4 stable was significantly older than the android-4.4 PELT, and that this patch was needed in android only because we forgot to backport it with the rest of the PELT patches. So yes, this one should go in the 4.4 stable upstream kernel.

Patch-set: 4
Reviewer: Gerrit User 1200908 <1200908@85c56323-6fa9-3386-8a01-6480fb634889>