UPSTREAM: arm64/lib: copy_page: use consistent prefetch stride
The optional prefetch instructions in the copy_page() routine are
inconsistent: at the start of the function, two cachelines are
prefetched beyond the one being loaded in the first iteration, but
in the loop, the prefetch is one more line ahead. This appears to
be unintentional, so let's fix it.

While at it, fix the comment style and white space.

Bug: 112277034
Change-Id: I5926b2d73fe7fc809e6595571e675bc562e76d7b
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Will Deacon <will.deacon@arm.com>
(cherry picked from commit 288be97cc74e31b7871c75eb11a8dd768dcb535d
Signed-off-by: Sami Tolvanen <samitolvanen@google.com>