ANDROID: staging: ashmem: Fix mmap size validation.
The ashmem driver did not check that the size/offset of the vma passed
to its .mmap() function was not larger than the ashmem object being
mapped. This could cause mmap() to succeed, even though accessing parts
of the mapping would later fail with a segmentation fault.

Ensure an error is returned by the ashmem_mmap() function if the vma
size is larger than the ashmem object size. This enables safer handling
of the problem in userspace.

Bug: 80147958
Change-Id: I128bd27d185f055b8adcd8a7cb0a63a121ef9f80
Signed-off-by: Alistair Strachan <astrachan@google.com>