Update patch set 2
Patch Set 2:

I think we ended up with something quite nice here, but we probably missed the window for it to go into android-4.9. I'm happy to leave this hanging in case we decide we need to resurrect it on 4.14 or later.

Patch-set: 2
Reviewer: Gerrit User 1122323 <1122323@85c56323-6fa9-3386-8a01-6480fb634889>