Update patch set 7
Patch Set 7:

> Patch Set 6:
> 
> (1 comment)
> 
> Here is another note on initialization.
> 
> Moreover, since we are at changing / improving the idle state tracking, I would also suggest to add an additional patch, maybe before this one, where we make more explicit that we are after the shallowest idle state.
> 
> What about renaming: best_idle_cstate in shallowest_idle_state?
> 
> This should make more clear that:
> - we initialize with the deepest possible idle C-State (INT_MAX)
> - we minimize the value of the idle state (i.e. pick the shallowest)

I agree that the the name of the variable could be more clear but I don't think it's reason enough to introduce another patch to this stack especially given that it would not bring any functional changes. Let's aim towards reviewing and merging this stack first and I'll add that to a "wish list" for further clean-up. By continuing to add patches to this stack we're making more difficult the review process.

Patch-set: 7
Reviewer: Gerrit User 1125654 <1125654@85c56323-6fa9-3386-8a01-6480fb634889>