Check Reverse-Mark Lookup Before ADDSA/DELSA
It's possible to insert an SA into the SADB that will
preclude the lookup of other SAs when using the MARK
attribute. The problem occurs based on a particular
sequencing with the marks where a new mark matches
requests for the SA mark of an existing SA but the
inverse is not true. For an example:
1) Add SA with mark=0, mask=0
2) Add an otherwise-identical SA
   with mark=0x1234, mask=0xFFFFFFFF

This will fail; however, if done in the reverse order
the second add will succeed:
1) Add an SA with mark=0x1234, mask=0xFFFFFFFF
2) Add an otherwise-identical SA
   except with mark=0, mask=0
Then:
3) Delete the SA using mark=0x1234, mask=0xFFFFFFF
4) Dump the SADB, and there will be one SA, and it will
have mark=0, mask=0.

This patch addresses the problem by performing a
reverse-match on the mark and preventing ADDSA for
any SA that would 'shadow' an existing SA in the SADB.

This patch also address a bug where it was possible to
add an SA with a mark greater than its mask, which could
never be deleted:
1) ADDSA with mark=0x1234, mask=0xFF
2) DELSA with mark=0x1234, mask=ZZZZ; error=ESRCH

Signed-off-by: Nathan Harold <nharold@google.com>