Update patch set 1
Patch Set 1:

Roman, We really, really don't want to add to our out-of-tree patch set in android common. I know you are working on upstreaming the goldfish patches... can we get that done first and backport them to android-4.9?

If there is heavy schedule pressure, then it might be better to have a goldfish branch downstream of android-4.9...

Patch-set: 1
CC: Gerrit User 1069543 <1069543@85c56323-6fa9-3386-8a01-6480fb634889>