Update patch set 1
Patch Set 1:

Is this patch not in 4.9 stable? If not, we should take it in through there, otherwise external folks (using the stable trees, not necessarily Android common) will also hit this issue.

Also, it should go in 4.4.  If we land it in upstream stable branch, they'll land in android/common branches as part of the merge backs.

Patch-set: 1
Reviewer: Gerrit User 1114194 <1114194@85c56323-6fa9-3386-8a01-6480fb634889>