Update patch set 4
Patch Set 4:

I think there's not much point to backporting to 3.18/4.4.

By my understanding:

launch kernel versions: Android X

3.x/3.11/3.18: Android M
3.11/3.18/4.4: Android N
3.18/4.4/4.9: Android O, oldest version that can be upgraded to Q
4.4/4.9/4.14: Android P
4.9/4.14/4.19: Android Q

There won't be any devices launching with Q (or even P) on 3.18.

Indeed the only way 3.18 will run Q is if a 3.18 launched on O device gets upgraded to P then Q.  During such an upgrade kernel won't (necessarily) be updated anyway, so won't pick up this change.

Or are we trying to backport this support into Pixel 1 (which IFIRC runs 3.18)?

Even if we do, it's not going to be something that we can rely on being present...
so userspace support will need to stay...

(by the way the same logic basically holds for the 4.4 backport of this patch, except wrt. Pixel 2 which IFIRC runs 4.4)

Patch-set: 4
Reviewer: Gerrit User 1003247 <1003247@85c56323-6fa9-3386-8a01-6480fb634889>