Update patch set 4
Patch Set 4:

> Or are we trying to backport this support into Pixel 1 (which IFIRC runs 3.18)?

Yeah, the idea is that we can test VTI support, and try to ensure that tunnel mode functionality works with Pixel 1 and 2.

> Even if we do, it's not going to be something that we can rely on being present...
> so userspace support will need to stay...
> 
> (by the way the same logic basically holds for the 4.4 backport of this patch, except wrt. Pixel 2 which IFIRC runs 4.4)

That is correct; userspace support will stay guarded by a feature flag, and this is all that is necessary to enable it on our two older devices.

Patch-set: 4
Reviewer: Gerrit User 1221725 <1221725@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0