Add comments clarifying the async model.
These comments annotate the functions with which locks are active or
acquired and relased.

It also moves the wireless_dev deallocation to the netdev destructor, to
fix some sequencing. nl80211 reacts to the netdev being deleted after
rtnetlink's dellink command, but before the netdev destructor. nl80211
wants to have the wireless_dev/wiphy reference when it reacts.

Bug: 116027983
Change-Id: Ie38252ce745abb89afce1ee533b0ecd0b30ebf97
Signed-off-by: A. Cody Schuffelen <schuffelen@google.com>
Acked-by: Greg Hartman <ghartman@google.com>
Acked-by: Tristan Muntsinger <muntsinger@google.com>