BACKPORT: kernel: Standardize vdso_datapage
In an effort to unify the common code for managing the vdso library in
between all the architectures that support it, this patch tries to
provide a common format for the vdso datapage.

As a result of this, this patch generalized the data structures in vgtod.h
from x86 private includes to general includes (include/vdso).

Cc: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>

(cherry picked from https://patchwork.kernel.org/patch/10729145/)
Bug: 20045882
Change-Id: I0600222971df3fba105cf434757fd33bbd627469