BACKPORT: kernel: Define gettimeofday vdso common code
In the last few years we assisted to an explosion of vdso
implementations that mostly share similar code.

This patch tries to unify the gettimeofday vdso implementation
introducing lib/vdso. The code contained in this library can
ideally be reused by all the architectures avoiding, where
possible, code duplication.

In doing so, tries to maintain the performances using inlining
as much as possible and consequently reduces the surface for
ROP type of attacks.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>

(cherry picked from https://patchwork.kernel.org/patch/10729155/)
Bug: 20045882
Change-Id: I04adedd0c7ef36d5a595f29e55996a9114691a5d