Update patch set 1
Patch Set 1:

> Patch Set 1:
> > Patch Set 1:
> > Is there a reason to do this as a separate patch? It's changing the same code as the previous patch so it seems better to just squash them together or drop this one.
> 
> ...especially since this is WALT-specific code that isn't going upstream or into android-4.19.

I just wanted to keep well separated the fix from a code refactoring. This makes easier to review and test boths: the first introduces a power/perf difference, while the second must not.

Moreover, this one is completely optional: if we want to limit changes in v4.14 to just fixes this should be probably dropped.

What do you prefer? If we want to add it I would still suggest to merge it separately, even just to keep the bugfix easy to read.

Patch-set: 1
Reviewer: Gerrit User 1119068 <1119068@85c56323-6fa9-3386-8a01-6480fb634889>