scsi: ufs: UFS driver v2.1 crypto support
Uses the UFSHCI v2.1 spec to manage keys in inline crypto engine
hardware, and exposes that functionality through the keyslot manager it
sets up in the device's request_queue. Uses the keyslot in the
bio_crypt_ctx of the bio, if specified, as the encryption context.

Known Issues: In the current implementation, multiple keyslot managers
may be allocated for a single UFS host. We should tie keyslot managers
to hosts to avoid this issue.

Change-Id: Icb133193f63749da540b282aa89250973fbd51da
Signed-off-by: Satya Tangirala <satyat@google.com>