Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Shouldn't a change like this go upstream first?
> 
> It's pretty much AOSP-specific, as androidkernel toolchains are used only here.

Android kernels are built by everyone, not only AOSP users :)

If this is a clang issue, it should go upstream.

If this is an issue of "a different tool should be used", then that too should go upstream as others will need the ability to select that as well.

There shouldn't be any aosp-only kernel patches unless there are very good reasons for it, as it is then required to be supported for 10-20 years by others.

Patch-set: 2
Reviewer: Gerrit User 1125653 <1125653@85c56323-6fa9-3386-8a01-6480fb634889>